loading data to all firms page and Firm Register UI

+++ src/main/java/com/lxisoft/service/FirmService.java
@@ -53,7 +53,6 @@ public interface FirmService {
      */
     void delete(Long id);
 
-    List<Firm> findFirmByCategory(int id);
-
+    List<Firm> findFirmByCategory(Long id);
 
 }
